Added JPrykop\File.pm
THIS IS A CODE SAMPLE from one of my longer-term projects;  I've swapped
out the name of that project with JPrykop in the code, but otherwise
made no changes.  Unfortunately I cannot release enough of the overall
system for it to compile, I can merely assure you that
it's been running smoothly for several years.